src: fix comparison of integer of different signedness

+++ src/libcrun/linux.c
@@ -1866,7 +1866,7 @@ libcrun_run_linux_container (libcrun_container_t *container,
 #define MAX_NAMESPACES 10
   cleanup_close_vec int *namespaces_to_join = (int[MAX_NAMESPACES+1]){-1};
   int namespaces_to_join_index[MAX_NAMESPACES];
-  int n_namespaces_to_join = 0;
+  size_t n_namespaces_to_join = 0;
   int userns_join_index = -1;
 
   for (i = 0; i < def->linux->namespaces_len; i++)
@@ -2298,7 +2298,7 @@ libcrun_join_process (libcrun_container_t *container, pid_t pid_to_join, libcrun
       ret = setns (fds[i], 0);
       if (UNLIKELY (ret < 0 && errno != EINVAL))
         {
-          int j;
+          size_t j;
           bool found = false;
 
           for (j = 0; j < def->linux->namespaces_len; j++)
